% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/threewords.R
\docType{package}
\name{threewords}
\alias{threewords}
\alias{threewords-package}
\title{Represent Precise Coordinates as Three Words}
\description{
This package provides a connector to \href{http://what3words.com/}{What3Words}, a database that represents 3mx3m
geographic areas as unique chains of 3 words
}

